Move GitHub Event to SwiftCIPlatforms
